member voting testing pass

+++ java/com/votingpoll/VotingSession.java
@@ -9,21 +9,21 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class VotingSession {
 
 	@Id
-	private Integer id;
+	private Long id;
 
 	private String name;
 	private Date startDate;
 	private Integer minutes;
 	
-	public VotingSession(Integer id, String name) {
+	public VotingSession(Long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
