+++ java/com/votingpoll/VotingSessionRepository.java
@@ -2,6 +2,6 @@ package com.votingpoll;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface VotingSessionRepository extends MongoRepository<VotingSession, Integer> {
+public interface VotingSessionRepository extends MongoRepository<VotingSession, Long> {
 
 }
